ashton v0.4.1 Ashton.Doc

Automatic opt documentation, to be placed into your function docstrings

Functions

document(schema, doc_opts \\ [])

document(schema :: Ashton.schema(), doc_opts :: Keyword.t()) :: String.t()

Opts

name(:string): The top level header for the opts documentation - Default: "Opts"

header_depth(:int): How many # to prepend before any heading - Default: 1

positive?(integer)