ashton v0.4.1 Ashton.SchemaHelpers

Helpers for building and working with schemas.

Functions

merge_schema()

merge_schema() :: Ashton.Schema.t()

The schema of the opts for merging two ashton schemas

schema_schema()

schema_schema() :: Ashton.Schema.t()

The schema of the opts for making an Ashton schema.